+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Logical folding for maven log in debug mode
 - Fold lines emitted directly to stdout and stderr (missing log level at start of line)
 - Fold indented debug lines
+- Fold error lines
 ### Changed
 - Improved Downloading/Progress/Downloaded log lines behavior
 ### Fixed